requirements.txt
llama-index==0.9.48
openai==1.3.0
python-dotenv==1.0.0